refactor: implementa service para validacao schema
criacao de usuarios
